fix #603; warning on automatic logout

+++ users/middleware.py
@@ -6,20 +6,24 @@ from ecs.users.models import UserProfile
 class SingleLogin(object):
     def process_request(self, request):
         """
-        If active, users can only have one active session
+        If active, users can only have one active session. This prevents
+        account sharing and opening the site in two browsers.
         """
+
         if not request.user.is_authenticated():
             return
 
-        session = Session.objects.get(session_key=request.session.session_key)
+        current_session = Session.objects.get(session_key=request.session.session_key)
         profile = request.user.ecs_profile
 
-        if profile.session_key != session.session_key:
-            old_session_key = profile.session_key
-            profile.session_key = session.session_key
-            profile.save()
+        if profile.session_key == current_session.session_key:
+            return
 
-            if old_session_key:
-                Session.objects.filter(session_key=old_session_key).delete()
+        old_session_key = profile.session_key
+        profile.session_key = current_session.session_key
+        profile.save()
 
+        if old_session_key and Session.objects.filter(session_key=old_session_key):
+            Session.objects.filter(session_key=old_session_key).delete()
+            request.single_login_enforced = True
